[refac] move dtos related with auth
