+++ src/main/java/org/hankki/hankkiserver/api/auth/controller/AuthController.java
@@ -7,9 +7,9 @@ import org.hankki.hankkiserver.common.ApiResponse;
 import org.hankki.hankkiserver.common.BaseResponse;
 import org.hankki.hankkiserver.common.code.SuccessCode;
 import org.hankki.hankkiserver.api.auth.service.AuthService;
-import org.hankki.hankkiserver.service.dto.request.UserLoginRequest;
-import org.hankki.hankkiserver.service.dto.response.UserLoginResponse;
-import org.hankki.hankkiserver.service.dto.response.UserReissueResponse;
+import org.hankki.hankkiserver.api.auth.controller.request.UserLoginRequest;
+import org.hankki.hankkiserver.api.auth.service.response.UserLoginResponse;
+import org.hankki.hankkiserver.api.auth.service.response.UserReissueResponse;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
